Add dynamite boom! & Add illumination at creation
Dynamite is instantiated in a non-lit state. Call LightFuse() to start
a countdown to the explosion. Note: the art is animated with a burning
fuse, despite the fact that it logically starts un-lit.

Checks for continued existance of illuminateable component on moving
objects. Objects can be killed mid-move (i.e. by dynamite).

Illuminateable components raise an event signifying their creation.
This is necessary so the OverWorld/CaveManager can update their
illumination to match the lightmap.